requirements.txt
numpy
matplotlib
Pillow
tkinter 